feat: cc-0077 hover change works for all courts
